feature(soundModeAndDifficultyAnalytics): reorganizing game files
